fix float again

+++ webui/webui.py
@@ -60,7 +60,7 @@ def format_aleo_credit(gates):
     return "{:,}".format(Decimal(gates) / 1_000_000)
 
 def format_number(number):
-    if isinstance(number, float):
+    if isinstance(number, Decimal):
         return f"{number:,.2f}"
     return "{:,}".format(number)
